// src/app/types/filters.types.ts
export type TBookingFilterRequest = {
    date?: Date | undefined;
    startTime?: Date | undefined;
    endTime?: Date | undefined;
    trainerName?: string | undefined;
};
export type TScheduleFilterRequest = {
    date?: Date | undefined;
    startTime?: Date | undefined;
    endTime?: Date | undefined;
    traineeName?: string | undefined;
    trainerName?: string | undefined;

};